+++ App.js
@@ -6,20 +6,20 @@ import { View } from 'react-native'
 import 'react-native-get-random-values'
 
 import { AgentProvider } from './App/contexts/AgentProvider'
-import Errors from './App/contexts/Errors/index'
-import Notifications from './App/contexts/Notifications/index'
+import Errors from './App/contexts/Errors'
+import Notifications from './App/contexts/Notifications'
 
-import EntryPoint from './App/screens/EntryPoint/index'
+import EntryPoint from './App/screens/EntryPoint'
 // import Home from './App/screens/Home/index'
 // import ListContacts from './App/screens/ListContacts/index'
 // import ListCredentials from './App/screens/ListCredentials/index'
 import Navbar from './App/components/Navbar'
-import PinCreate from './App/screens/PinCreate/index'
-import PinEnter from './App/screens/PinEnter/index'
+import PinCreate from './App/screens/PinCreate'
+import PinEnter from './App/screens/PinEnter'
 // import Settings from './App/screens/Settings/index'
-import SetupWizard from './App/screens/SetupWizard/index'
+import SetupWizard from './App/screens/SetupWizard'
 import Terms from './App/components/Terms'
-import Workflow from './App/screens/Workflow/index'
+import Workflow from './App/screens/Workflow'
 
 import TabNavigator from './App/navigators/TabNavigator'
