--- App/screens/ListContacts/styles.js
@@ -1,28 +1,0 @@
-import React from 'react'
-
-import { StyleSheet } from 'react-native'
-
-const styles = StyleSheet.create({
-  backbutton: {
-    marginBottom: 30,
-  },
-  tableItem: {
-    paddingLeft: 30,
-    display: 'flex',
-    alignItems: 'center',
-    borderBottomWidth: 1,
-    borderColor: '#fff',
-  },
-  tableSubItem: {
-    height: 50,
-  },
-  credView: {
-    alignItems: 'center',
-    padding: 12,
-    borderTopLeftRadius: 50,
-    borderTopRightRadius: 50,
-    height: '100%',
-  },
-})
-
-export default styles
