+++ App/screens/Offered.tsx
@@ -1,6 +1,6 @@
-import React, { useState, useContext } from 'react'
+import React, { useState, useEffect, useContext } from 'react'
 
-import { Image, Text, TouchableOpacity, View } from 'react-native'
+import { Image, Text, TouchableOpacity, View, StyleSheet } from 'react-native'
 
 import { useHistory } from 'react-router-native'
 
@@ -9,28 +9,39 @@ import BackButton from '../../../../components/BackButton/index'
 import CurrentContact from '../../../../components/CurrentContact/index'
 import CurrentCredential from '../../../../components/CurrentCredential/index'
 
-import { ErrorsContext } from '../../../../contexts/Errors/index'
-import { NotificationsContext } from '../../../../contexts/Notifications/index'
+import AgentContext from '../contexts/AgentProvider'
 
-import Styles from '../styles'
-import Images from '../../../../../assets/images'
-import AppStyles from '../../../../../assets/styles'
-import { IContact, ICredential } from '../../../../../types'
+import Images from '../../assets/images'
+import AppStyles from '../../assets/styles'
+import { IContact, ICredential } from '../../types'
 
-interface ICredentialRequested {
+interface ICredentialOffered {
   contact: IContact
   credential: ICredential
 }
 
-function CredentialRequested(props: ICredentialRequested) {
+function CredentialOffered(props: ICredentialOffered) {
   const history = useHistory()
 
-  const errors = useContext(ErrorsContext)
-  const notifications = useContext(NotificationsContext)
+  //Reference to the agent context
+  const agentContext = useContext(AgentContext)
+
+  useEffect(() => {
+    console.log(props.credential)
+  }, [])
 
   const [viewInfo, setViewInfo] = useState('')
   const [viewContact, setViewContact] = useState(false)
   const [viewCredential, setViewCredential] = useState(false)
+
+  const acceptOffer = async () => {
+    console.log('Attempting to accept offer')
+    await agentContext.agent.credentials.acceptOffer(props.credential.id)
+
+    //TODO:
+    //Push to pending issuance screen
+    history.push('/workflow/pending')
+  }
 
   return (
     <>
@@ -40,14 +51,14 @@ function CredentialRequested(props: ICredentialRequested) {
           <View style={AppStyles.header}>
             <AppHeader headerText={'CREDENTIALS'} />
           </View>
-          <View style={[AppStyles.tab, Styles.tabView]}>
+          <View style={[AppStyles.tab, styles.tabView]}>
             <Text style={[AppStyles.h3, AppStyles.textSecondary, AppStyles.textUpper, AppStyles.textBold]}>
-              Connected to:
+              Credential from:
             </Text>
             <View style={AppStyles.tableItem}>
               <View>
-                <Text style={[{ fontSize: 18 }, AppStyles.textSecondary, AppStyles.textUpper, AppStyles.textBold]}>
-                  {props.contact.label}
+                <Text style={[{ fontSize: 20, top: 8 }, AppStyles.textBlack, AppStyles.textBold]}>
+                  {props.contact.alias ? props.contact.alias : props.contact.invitation.label}
                 </Text>
                 <Text style={[{ fontSize: 14 }, AppStyles.textSecondary]}>{props.contact.sublabel}</Text>
               </View>
@@ -62,10 +73,10 @@ function CredentialRequested(props: ICredentialRequested) {
             </View>
             <View style={AppStyles.tableItem}>
               <View>
-                <Text style={[{ fontSize: 18 }, AppStyles.textSecondary, AppStyles.textUpper, AppStyles.textBold]}>
-                  {props.credential.label}
+                <Text style={[{ fontSize: 20, top: 8 }, AppStyles.textSecondary, AppStyles.textBold]}>
+                  {/*Temporary hardcoding*/}
+                  Driver's License
                 </Text>
-                <Text style={[{ fontSize: 14 }, AppStyles.textSecondary]}>{props.credential.sublabel}</Text>
               </View>
               <TouchableOpacity
                 onPress={() => {
@@ -76,7 +87,7 @@ function CredentialRequested(props: ICredentialRequested) {
                 <Image source={Images.infoGray} style={[AppStyles.info, { marginRight: 0, top: 10 }]} />
               </TouchableOpacity>
             </View>
-            <View style={Styles.buttonWrap}>
+            <View style={styles.buttonWrap}>
               <Text
                 style={[
                   { fontSize: 18 },
@@ -86,15 +97,15 @@ function CredentialRequested(props: ICredentialRequested) {
                   AppStyles.textBold,
                 ]}
               >
-                Send Credentials
+                Claim Credentials
               </Text>
               <TouchableOpacity
-                style={[Styles.button, AppStyles.backgroundPrimary]}
+                style={[styles.button, AppStyles.backgroundPrimary]}
                 onPress={() => {
-                  history.push('/home')
+                  acceptOffer()
                 }}
               >
-                <Image source={Images.send} style={Styles.buttonIcon} />
+                <Image source={Images.receive} style={styles.buttonIcon} />
               </TouchableOpacity>
             </View>
             <TouchableOpacity
@@ -103,7 +114,7 @@ function CredentialRequested(props: ICredentialRequested) {
                 history.push('/home')
               }}
             >
-              <Text style={[{ fontSize: 14 }, AppStyles.textGray, AppStyles.textCenter]}>Decline{'\n'}Request</Text>
+              <Text style={[{ fontSize: 14 }, AppStyles.textGray, AppStyles.textCenter]}>Decline{'\n'}Offer</Text>
             </TouchableOpacity>
           </View>
         </View>
@@ -114,4 +125,42 @@ function CredentialRequested(props: ICredentialRequested) {
   )
 }
 
-export default CredentialRequested
+export default CredentialOffered
+
+const styles = StyleSheet.create({
+  tabView: {
+    backgroundColor: '#ddd',
+    borderTopLeftRadius: 50,
+    borderTopRightRadius: 50,
+    padding: '7%',
+  },
+  icon: {
+    fontSize: 26,
+    top: 10,
+  },
+  button: {
+    marginVertical: 7,
+    width: 125,
+    paddingVertical: 10,
+    borderRadius: 20,
+    alignItems: 'center',
+  },
+  buttonText: {
+    maxWidth: 120,
+    flexWrap: 'wrap',
+    textAlign: 'right',
+    marginVertical: 10,
+    marginHorizontal: 20,
+  },
+  buttonWrap: {
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    left: -10,
+    top: 40,
+  },
+  buttonIcon: {
+    width: 30,
+    height: 43,
+  },
+})
