+++ App/screens/PinEnter.tsx
@@ -6,10 +6,10 @@ import { useHistory } from 'react-router-native'
 
 import * as Keychain from 'react-native-keychain'
 
-import AppHeaderLarge from '../../components/AppHeaderLarge'
-import LoadingOverlay from '../../components/LoadingOverlay'
+import AppHeaderLarge from '../components/AppHeaderLarge'
+import LoadingOverlay from '../components/LoadingOverlay'
 
-import AppStyles from '../../../assets/styles'
+import AppStyles from '../../assets/styles'
 
 function PinEnter(props) {
   const history = useHistory()
