--- App/screens/Settings/styles.js
@@ -1,14 +1,0 @@
-import React from 'react'
-
-import { StyleSheet } from 'react-native'
-
-const styles = StyleSheet.create({
-  backbutton: {
-    marginBottom: 30,
-  },
-  settingView: {
-    alignItems: 'center',
-  },
-})
-
-export default styles
